Update 1.4
Now it shows when the pause and the replay buttons are activated. It shows when the volume is muted or not. If the user tries to add a song that already exists, the app shows them that this song already exists and it asks them if they want to keep it in the playlist. New colors to the GUI and prety much this is the final build